feat : BaseResponse 구현
- 클라이언트에게 반환할 최상위 HTTP Response 클래스 작성
- 디렉터리 구조 일부 변경
