+++ src/main/java/com/example/demo/domain/dto/request/RegisterBookDTO.java
@@ -1,0 +1,16 @@
+package com.example.demo.domain.dto.request;
+
+import com.example.demo.domain.entity.BookCategory;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterBookDTO {
+    String title;
+    String isbn;
+    String author;
+    BookCategory category;
+}
