+++ src/main/java/com/example/demo/domain/dto/response/BaseResponse.java
@@ -1,0 +1,10 @@
+package com.example.demo.domain.dto.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class BaseResponse {
+    private String message;
+}
