+++ src/main/java/com/example/demo/domain/dto/response/SimpleBookResponse.java
@@ -1,0 +1,14 @@
+package com.example.demo.domain.dto.response;
+
+import com.example.demo.domain.entity.BookCategory;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class SimpleBookResponse extends BaseResponse{
+    String title;
+    String isbn;
+    String author;
+    BookCategory category;
+}
